first board drawn

+++ proj/src/game/objects/board/board.h
@@ -10,6 +10,6 @@ Board table[BOARD_SIZE][BOARD_SIZE];
 
 
 // returns the color of the square in lin/col
-int get_square_color(uint8_t lin, uint8_t col);
+int get_square_color(uint16_t lin, uint16_t col);
 
 #endif
